% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{add_symmetry_axis}
\alias{add_symmetry_axis}
\title{Append a symmetry axis to a Molecule3D}
\usage{
add_symmetry_axis(molecule, symmetry_axis, label = "Symmetry Axis")
}
\arguments{
\item{molecule}{A \code{\link{Molecule3D}} object.}

\item{symmetry_axis}{A \code{\link{SymAxis}} object to append.}
}
\value{
A \code{Molecule3D} object with \code{symmetry_axis} appended to
\code{@symmetry_axes}.
}
\description{
Adds a single \code{\link{SymAxis}} object to the \code{@symmetry_axes} list
of a \code{\link{Molecule3D}}. The axis is appended (order preserved).
}
\examples{
atoms <- data.frame(
  eleno = c(1, 2),
  elena = c("C","O"),
  x = c(0, 1), y = c(0, 0), z = c(0, 0)
)
bonds <- data.frame(bond_id = 1, origin_atom_id = 1, target_atom_id = 2)
m  <- Molecule3D("CO", atoms = atoms, bonds = bonds)
ax <- SymAxis(Cn = 2L, posA = c(0,0,0), posB = c(0,0,1))
m  <- add_symmetry_axis(m, ax)
length(m@symmetry_axes)  # 1
m@symmetry_axes_orders   # 2

}
\seealso{
\link{SymAxis}, \code{\link[=is_symmetry_axis]{is_symmetry_axis()}},
\code{\link[=fetch_all_symmetry_axes_with_order]{fetch_all_symmetry_axes_with_order()}}, \code{\link[=transform_symmetry_axis]{transform_symmetry_axis()}}
}
