% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-SymmetryElements.R
\name{add_symmetry_element_to_collection}
\alias{add_symmetry_element_to_collection}
\title{Add a symmetry element to a collection}
\usage{
add_symmetry_element_to_collection(collection, new)
}
\arguments{
\item{collection}{A \code{\link{SymmetryElementCollection}} object.}

\item{new}{A \code{\link{SymmetryElement}} to append to \code{collection}.}
}
\value{
A \code{\link{SymmetryElementCollection}} with \code{new} appended and
\code{ids} updated to include the newly assigned ID.
}
\description{
Appends a single \code{\link{SymmetryElement}} to an existing
\code{\link{SymmetryElementCollection}}, assigning it the next available numeric ID.
}
\examples{
mp <- MirrorPlane(normal = c(0, 0, 1), position = c(0, 0, 0), label = "<U+03C3>_xy")
coll <- SymmetryElementCollection()
coll2 <- add_symmetry_element_to_collection(coll, mp)
coll2@ids # contains 1

}
