% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{add_symmetry_element_to_molecule}
\alias{add_symmetry_element_to_molecule}
\title{Append a symmetry axis to a Molecule3D}
\usage{
add_symmetry_element_to_molecule(molecule, symmetry_element)
}
\arguments{
\item{molecule}{A \code{\link{Molecule3D}} object.}

\item{symmetry_element}{A \code{\link{SymmetryElement}} object to append.}
}
\value{
A \code{Molecule3D} object with \code{symmetry_element} appended to
\code{@symmetry_elements}.
}
\description{
Adds a single \code{\link{SymmetryElement}} object to the \code{@symmetry_elements} paramater
of a \code{\link{Molecule3D}}. The element is appended (order preserved) and given a unique ID.
}
\examples{
atoms <- data.frame(
  eleno = c(1, 2),
  elena = c("C", "O"),
  x = c(0, 1), y = c(0, 0), z = c(0, 0)
)
bonds <- data.frame(bond_id = 1, origin_atom_id = 1, target_atom_id = 2)
m <- Molecule3D("CO", atoms = atoms, bonds = bonds)
ax <- ProperRotationAxis(n = 2, posA = c(0, 0, 0), posB = c(0, 0, 1))
m <- add_proper_rotation_axis_to_molecule(m, ax)

length(m@symmetry_elements@proper_rotation_axes) # 1
m@symmetry_elements@unique_proper_axis_orders # 2

}
