% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-SymmetryElements.R
\name{CentreOfInversion}
\alias{CentreOfInversion}
\title{CentreOfInversion: inversion symmetry element (i)}
\usage{
CentreOfInversion(position, label = "unnamed")
}
\arguments{
\item{position}{Numeric length-3 vector giving the inversion centre
(\code{c(x, y, z)}).}

\item{label}{Character scalar. Optional user-defined name or description
for the symmetry element.}
}
\value{
An S7 object of class \code{CentreOfInversion}.
}
\description{
Represents a centre of inversion in 3D space. Every point \code{p} is mapped to
\verb{p' = 2 * position - p} under inversion through this centre.
}
\examples{
# Inversion centre at the origin
ci <- CentreOfInversion(
  position = c(0, 0, 0),
  label = "i"
)

ci@type    # "Centre of Inversion"

}
