% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{compute_plane_from_atoms}
\alias{compute_plane_from_atoms}
\title{Compute a best-fit plane through all atoms of a Molecule3D}
\usage{
compute_plane_from_atoms(molecule)
}
\arguments{
\item{molecule}{A \code{\link{Molecule3D}} object.}
}
\value{
A list describing the plane (normal & offset). See \code{\link[move:compute_plane_from_points]{move::compute_plane_from_points()}} for details,
}
\description{
Extracts atomic Cartesian coordinates from a \code{\link{Molecule3D}}
object and computes the best-fit plane using
\code{\link[move:compute_plane_from_points]{move::compute_plane_from_points()}}. If only 3 atoms are present, it computes an exact solution.
If there are more than 3 points it computes the best-fit plane by SVD.
}
\examples{
m <- read_mol2(system.file("benzene.mol2", package = "structures"))
plane <- compute_plane_from_atoms(m)
plane$normal

}
