% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{fetch_all_proper_rotation_axes_with_order}
\alias{fetch_all_proper_rotation_axes_with_order}
\title{Fetch all proper rotation axes of a given order (Cn) from a Molecule3D}
\usage{
fetch_all_proper_rotation_axes_with_order(molecule, Cn)
}
\arguments{
\item{molecule}{A \code{\link{Molecule3D}} object.}

\item{Cn}{Integer (or integer-like numeric) fold/order to match
(e.g. \code{2}, \code{3}, \code{6}).}
}
\value{
\itemize{
\item \code{NULL} if the molecule contains no symmetry elements at all.
\item Otherwise, a \emph{list} of \code{\link{ProperRotationAxis}}
objects of order \code{Cn}. The list may be empty if no such axes
exist.
}
}
\description{
Convenience wrapper around
\code{\link[=fetch_all_proper_rotation_axes_with_order_from_collection]{fetch_all_proper_rotation_axes_with_order_from_collection()}} that operates
directly on a \code{\link{Molecule3D}} by using its
\code{@symmetry_elements} collection.
}
\examples{
atoms <- data.frame(
  eleno = c(1, 2),
  elena = c("C","O"),
  x = c(0, 1), y = c(0, 0), z = c(0, 0)
)
bonds <- data.frame(
  bond_id = 1,
  origin_atom_id = 1,
  target_atom_id = 2
)
m <- Molecule3D("CO", atoms = atoms, bonds = bonds)

m <- add_symmetry_element_to_molecule(
  m,
  ProperRotationAxis(n = 2L, posA = c(0, 0, 0), posB = c(0, 0, 1), label = "C2(z)")
)

fetch_all_proper_rotation_axes_with_order(m, Cn = 2L)

}
