% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{fetch_eleno_by_atom_type}
\alias{fetch_eleno_by_atom_type}
\title{Fetch atom identifiers by SYBYL atom type}
\usage{
fetch_eleno_by_atom_type(x, atom_type)
}
\arguments{
\item{x}{A \code{\link{Molecule3D}} object.}

\item{atom_type}{Character vector of SYBYL atom types to match
(values in \code{x@atoms$atom_type}; see \code{\link[=valid_atom_types]{valid_atom_types()}}).}
}
\value{
A numeric vector of atom identifiers (\code{eleno}) whose
\code{atom_type} matches any value in \code{atom_type}.
}
\description{
Fetch atom identifiers by SYBYL atom type
}
\examples{
path <- system.file(package = "structures", "fe_dummies.mol2")
molecule <- read_mol2(path)

# Get Iron and dummy atoms
fetch_eleno_by_atom_type(molecule, c("Fe", "Du", "Du.C"))
}
