% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{fetch_eleno_by_element}
\alias{fetch_eleno_by_element}
\title{Fetch atom identifiers by element}
\usage{
fetch_eleno_by_element(x, element)
}
\arguments{
\item{x}{a Molecule3D object}

\item{element}{element name}
}
\value{
a numeric vector of atom identifiers (eleno) corresponding to element names (elena).
}
\description{
Fetch atom identifiers by element
}
\examples{
path <- system.file(package="structures", "benzene.mol2")
molecule <- read_mol2(path)
fetch_eleno_by_element(molecule, "C")

}
