% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{fetch_eleno_connected_by_bond}
\alias{fetch_eleno_connected_by_bond}
\title{Fetch the two atoms connected by a bond}
\usage{
fetch_eleno_connected_by_bond(molecule, bond_id)
}
\arguments{
\item{molecule}{A \code{Molecule3D} object.}

\item{bond_id}{Numeric bond identifier (must exist in \code{molecule@bond_ids}).}
}
\value{
A numeric vector of length 2 giving the \code{origin_atom_id} and
\code{target_atom_id} for the specified bond.
}
\description{
Returns the numeric atom IDs (\code{eleno}) of the two atoms joined by a given
bond in a \code{Molecule3D} object. This is a convenience helper to quickly
inspect bond connectivity or to use as input for other geometric operations.
}
\examples{
m <- read_mol2(system.file("fe_tripod.mol2", package = "structures"))
fetch_eleno_connected_by_bond(m, bond_id = 16)

}
\seealso{
\code{\link{fetch_eleno_downstream_of_bond}}, \code{\link{add_bonds}}
}
