% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{fetch_symmetry_axis_by_id}
\alias{fetch_symmetry_axis_by_id}
\title{Fetch a single symmetry axis by ID}
\usage{
fetch_symmetry_axis_by_id(molecule, id)
}
\arguments{
\item{molecule}{A \code{\link[=Molecule3D]{Molecule3D()}} object.}

\item{id}{the ID of a symmetry axis. See \verb{@symmetry_axes_dataframe} property for a description of all available asymmetry axes}
}
\value{
A \code{\link[=ProperRotationAxis]{ProperRotationAxis()}} object
}
\description{
Fetch a single symmetry axis by ID
}
\examples{
# Minimal molecule
atoms <- data.frame(
  eleno = c(1, 2),
  elena = c("C", "O"),
  x = c(0, 1), y = c(0, 0), z = c(0, 0)
)
bonds <- data.frame(
  bond_id = 1,
  origin_atom_id = 1,
  target_atom_id = 2
)
m <- Molecule3D("CO", atoms = atoms, bonds = bonds)

# Append a symmetry axis (ID is assigned automatically)
m <- add_proper_rotation_axis(m, ProperRotationAxis(Cn = 2L, posA = c(0, 0, 0), posB = c(0, 0, 1)))

# Inspect available axis IDs
m@symmetry_axes_dataframe$id

# Fetch the first axis by its ID
ax_id <- m@symmetry_axes_dataframe$id[1]
ax <- fetch_symmetry_axis_by_id(m, ax_id)
ax
}
