% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{fetch_symmetry_element_from_molecule}
\alias{fetch_symmetry_element_from_molecule}
\title{Fetch a single symmetry element by ID}
\usage{
fetch_symmetry_element_from_molecule(molecule, id, error_if_missing = TRUE)
}
\arguments{
\item{molecule}{A \code{\link[=Molecule3D]{Molecule3D()}} object.}

\item{id}{the ID of a symmetry element.}

\item{error_if_missing}{Logical; if \code{TRUE} (default), an error is thrown
when \code{id} is not found in the collection. If \code{FALSE}, the
function returns \code{NULL} when \code{id} is missing.}
}
\value{
A \code{\link[=ProperRotationAxis]{ProperRotationAxis()}} object
}
\description{
Fetch a single symmetry element by ID
}
\examples{
# Minimal molecule
atoms <- data.frame(
  eleno = c(1, 2),
  elena = c("C", "O"),
  x = c(0, 1), y = c(0, 0), z = c(0, 0)
)
bonds <- data.frame(
  bond_id = 1,
  origin_atom_id = 1,
  target_atom_id = 2
)
m <- Molecule3D("CO", atoms = atoms, bonds = bonds)

# Append a symmetry axis (ID is assigned automatically)
m <- add_symmetry_element_to_molecule(
    m,
    ProperRotationAxis(n = 2, posA = c(0, 0, 0), posB = c(0, 0, 1))
  )

# Inspect available axis IDs
m@symmetry_elements@summary

# Fetch the first axis by its ID
ax_id <- m@symmetry_elements@summary$ids[1]
ax <- fetch_symmetry_element_from_molecule(m, ax_id)
ax
}
