% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{filter_atoms}
\alias{filter_atoms}
\title{Filter molecule for specific atoms}
\usage{
filter_atoms(x, eleno)
}
\arguments{
\item{x}{a Molecule3D object}

\item{eleno}{a vector of atom IDs to filter. Will automatically cast to character vector.}
}
\value{
a Molecule3D object
}
\description{
Filters a molecule to include only specific atoms by atom_id (eleno).
Will also drop orphaned bonds. To delete atoms from a molecule, see \code{\link[=remove_atoms]{remove_atoms()}}
}
\examples{
path <- system.file(package="structures", "benzene.mol2")
molecule <- read_mol2(path)
filter_atoms(molecule, c(1,2,3))
}
