% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-SymmetryElements.R
\name{is_centre_of_inversion}
\alias{is_centre_of_inversion}
\title{Check if an object is a CentreOfInversion}
\usage{
is_centre_of_inversion(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical. \code{TRUE} if \code{x} inherits from \code{\link{CentreOfInversion}}, otherwise \code{FALSE}.
}
\description{
Check if an object is a CentreOfInversion
}
\examples{
ci <- CentreOfInversion(position = c(0, 0, 0))
is_centre_of_inversion(ci) # TRUE
is_centre_of_inversion(123) # FALSE

}
