% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-SymmetryElements.R
\name{is_improper_rotation_axis}
\alias{is_improper_rotation_axis}
\title{Check if an object is an ImproperRotationAxis}
\usage{
is_improper_rotation_axis(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical. \code{TRUE} if \code{x} inherits from \code{\link{ImproperRotationAxis}}, otherwise \code{FALSE}.
}
\description{
Check if an object is an ImproperRotationAxis
}
\examples{
# Assuming ImproperRotationAxis() will have a constructor similar to other elements:
iax <- ImproperRotationAxis()
is_improper_rotation_axis(iax)   # TRUE
is_improper_rotation_axis(NULL)  # FALSE

}
