% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-SymmetryElements.R
\name{is_mirror_plane}
\alias{is_mirror_plane}
\title{Check if an object is a MirrorPlane}
\usage{
is_mirror_plane(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical. \code{TRUE} if \code{x} inherits from \code{\link{MirrorPlane}}, otherwise \code{FALSE}.
}
\description{
Check if an object is a MirrorPlane
}
\examples{
mp <- MirrorPlane(normal = c(0,0,1), position = c(0,0,0))
is_mirror_plane(mp)        # TRUE
is_mirror_plane("nope")    # FALSE

}
