% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{is_molecule}
\alias{is_molecule}
\title{Check if an object is a Molecule3D}
\usage{
is_molecule(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical scalar: \code{TRUE} if \code{x} is a \code{Molecule3D}, otherwise \code{FALSE}.
}
\description{
Tests whether an object inherits from the \code{\link{Molecule3D}} class.
}
\examples{
atoms <- data.frame(eleno = c(1,2), elena = c("C","O"), x=c(0,1), y=c(0,0), z=c(0,0))
bonds <- data.frame(bond_id = 1, origin_atom_id = 1, target_atom_id = 2)
mol <- Molecule3D(name = "CO", atoms = atoms, bonds = bonds)
is_molecule(mol)
is_molecule(123)

}
