% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-SymmetryElements.R
\name{is_proper_rotation_axis}
\alias{is_proper_rotation_axis}
\title{Check if an object is a ProperRotationAxis}
\usage{
is_proper_rotation_axis(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical. \code{TRUE} if \code{x} inherits from \code{\link{ProperRotationAxis}}
(and not from \code{ImproperRotationAxis}), otherwise \code{FALSE}.
}
\description{
Check if an object is a ProperRotationAxis
}
\examples{
ax <- ProperRotationAxis(Cn = 3L, posA = c(0,0,0), posB = c(0,0,1))
is_proper_rotation_axis(ax)        # TRUE
is_proper_rotation_axis("nope")    # FALSE

}
