% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-SymmetryElements.R
\name{is_symmetry_axis}
\alias{is_symmetry_axis}
\title{Check if an object is a symmetry axis}
\usage{
is_symmetry_axis(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
A logical scalar: \code{TRUE} if \code{x} is a \code{ProperRotationAxis}, otherwise \code{FALSE}.
}
\description{
Tests whether an object inherits from the \code{\link{ProperRotationAxis}} or \code{\link{ImproperRotationAxis}} class.
}
\examples{
ax <- ProperRotationAxis(n = 2L, posA = c(0,0,0), posB = c(0,0,1))
is_symmetry_axis(ax)
is_symmetry_axis("not_an_axis")

}
