% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{minimal_bonds}
\alias{minimal_bonds}
\title{Create a Minimal Bonds Data Frame}
\usage{
minimal_bonds()
}
\value{
A \code{data.frame} with columns:
\describe{
\item{bond_id}{Numeric vector of bond identifiers.}
\item{origin_atom_id}{Character vector of origin atom IDs (by eleno).}
\item{target_atom_id}{Character vector of target atom IDs (by eleno).}
}
Each column is initialized as a zero-length vector.
}
\description{
Generates an empty \code{data.frame} containing the required columns for bonds
in a \code{Molecule3D} object. This serves as the default \code{bonds} input when
creating a new \code{Molecule3D}.
}
\examples{
minimal_bonds()

}
