% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointGroup.R
\name{PointGroup3D}
\alias{PointGroup3D}
\title{PointGroup3D: A 3D Point Group}
\usage{
PointGroup3D(name = "", symmetry_elements = list())
}
\arguments{
\item{name}{String (length 1). What is the name of this point group. We do not restrict this since with no lattice restriction there are infinitely many point groups.
However, if you are creating one of the 32 Crystallographic 3D point groups (the ones which allow for periodic crystals) we recommend setting this based on the Schönflies notation. See \code{\link[=crystallographic_pointgroup_names]{crystallographic_pointgroup_names()}}.}

\item{symmetry_elements}{A list of \code{\link[=SymmetryElement]{SymmetryElement()}} objects.}
}
\description{
In geometry, a point group is a list of Symmetry Elements (see \code{\link[=SymmetryElement]{SymmetryElement()}}) about
which symmetry operations can be performed without changing the position of atoms.
}
