% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse.R
\name{read_mol2}
\alias{read_mol2}
\title{Read Mol2 file to a 'Structure' class object}
\usage{
read_mol2(path)
}
\arguments{
\item{path}{path to the mol2 file}
}
\value{
a Structure class object
}
\description{
Read Mol2 file to a 'Structure' class object
}
\examples{
mol2 <- system.file(package="structures", "benzene.mol2")
read_mol2(mol2)
}
