% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{remove_atoms}
\alias{remove_atoms}
\title{Delete atoms from molecule}
\usage{
remove_atoms(x, eleno)
}
\arguments{
\item{x}{a Molecule3D object}

\item{eleno}{a vector of atom IDs to delete}
}
\value{
a Molecule3D object
}
\description{
Removes atoms from molecule by atom_id (eleno).
Will also drop bonds that involve deleted atoms.
}
\examples{
path <- system.file(package="structures", "benzene.mol2")
molecule <- read_mol2(path)
remove_atoms(molecule, c(1,2,3))
}
