% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{remove_dummy_atoms}
\alias{remove_dummy_atoms}
\title{Remove Dummy atoms}
\usage{
remove_dummy_atoms(x, dummy_type = c("Du", "Du.C"))
}
\arguments{
\item{x}{a \code{\link[=Molecule3D]{Molecule3D()}} object}
}
\value{
a \code{\link[=Molecule3D]{Molecule3D()}} object with dummy atoms (and any bonds involving dummy atoms) removed
}
\description{
Remove dummy atoms from a molecule.
Dummy are identified based on 'atom_type' (according to SYBYL mol2 specification should be set to 'Du' / 'Du.C' for dummy atoms).
}
\examples{
path <- system.file(package = "structures", "fe_dummies.mol2")
molecule_with_dummies <- read_mol2(path)
remove_dummy_atoms(molecule_with_dummies)
}
