% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{set_anchor_by_atom}
\alias{set_anchor_by_atom}
\title{Set the molecule anchor using an atom ID}
\usage{
set_anchor_by_atom(x, eleno)
}
\arguments{
\item{x}{A \code{Molecule3D} object.}

\item{eleno}{Atom ID (matching \code{x@atoms$eleno}).}
}
\value{
A \code{Molecule3D} object with \code{x@anchor} set to the atom position.
}
\description{
Sets the \code{anchor} of a \code{Molecule3D} to the current coordinates
of a specified atom (\code{eleno}).
}
\examples{
# m <- set_anchor_by_atom(m, "5")
}
\seealso{
\code{\link{set_anchor_by_position}},
\code{\link{translate_molecule_to_origin}},
\code{\link{translate_molecule_to_position}},
\code{\link{translate_molecule_by_vector}}
}
