% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{set_anchor_by_position}
\alias{set_anchor_by_position}
\title{Set the molecule anchor by position}
\usage{
set_anchor_by_position(x, position)
}
\arguments{
\item{x}{A \code{Molecule3D} object.}

\item{position}{Numeric length-3 vector \code{c(x, y, z)}.}
}
\value{
A \code{Molecule3D} object with \code{x@anchor} set to \code{position}.
}
\description{
Sets the \code{anchor} of a \code{Molecule3D} to an arbitrary position
\code{c(x, y, z)}. The anchor is a reference point used by translation
helpers to reposition the molecule.
}
\examples{
# m <- set_anchor_by_position(m, c(1, 2, 3))
}
\seealso{
\code{\link{set_anchor_by_atom}},
\code{\link{translate_molecule_to_origin}},
\code{\link{translate_molecule_to_position}},
\code{\link{translate_molecule_by_vector}}
}
