% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapes.R
\name{shape_cube}
\alias{shape_cube}
\title{Cube (regular)}
\usage{
shape_cube()
}
\value{
A \code{\link{Shape}} representing a cube.
}
\description{
Constructs a unit cube centered at the origin with vertices at all
combinations of \verb{(±1, ±1, ±1)}. Edges are determined by unit Hamming-distance.
}
\examples{
shp <- shape_cube()
shp

}
