% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shapes.R
\name{shape_dodecahedron}
\alias{shape_dodecahedron}
\title{Dodecahedron (regular)}
\usage{
shape_dodecahedron()
}
\value{
A \code{\link{Shape}} representing a dodecahedron.
}
\description{
Constructs a regular dodecahedron using golden ratio relationships, combining
cube corners and additional vertices. Edges are derived from minimal equal
edge lengths.
}
\examples{
shp <- shape_dodecahedron()
shp

}
