% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetries.R
\name{shape_octahedron}
\alias{shape_octahedron}
\title{Octahedron (regular)}
\usage{
shape_octahedron()
}
\value{
A \code{\link{Shape}} representing an octahedron.
}
\description{
Constructs a regular octahedron centered at the origin with vertices on
the coordinate axes \verb{(±1, 0, 0)}, \verb{(0, ±1, 0)}, \verb{(0, 0, ±1)}.
}
\examples{
shp <- shape_octahedron()
shp

}
