% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetries.R
\name{shape_tetrahedron}
\alias{shape_tetrahedron}
\title{Tetrahedron (regular)}
\usage{
shape_tetrahedron()
}
\value{
A \code{\link{Shape}} representing a tetrahedron.
}
\description{
Constructs a regular tetrahedron centered at the origin with vertices at
permutations of \verb{(±1, ±1, ±1)} of odd parity. Edges and faces are populated.
}
\examples{
shp <- shape_tetrahedron()
shp

}
