% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{Structure}
\alias{Structure}
\title{Create Structure Object}
\usage{
Structure(
  atoms = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names))
    {
         list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })(),
  bonds = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names))
    {
         list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })()
)
}
\arguments{
\item{atoms}{data.frame of atoms with columns:  eleno (element number) and elena (element name)}

\item{bonds}{data.frame of bonds (1 row per bond) with columns: from & to where values represent eleno of the atoms they are connecting. Direction is meaningless}
}
\value{
an object of class 'Structure'
}
\description{
Create Structure Object
}
