% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetries.R
\name{symmetry_tetrahedron}
\alias{symmetry_tetrahedron}
\title{Symmetry Tetrahedron}
\usage{
symmetry_tetrahedron()
}
\value{
a list of symmetry relevant information
}
\description{
Symmetry Tetrahedron
}
\examples{
symmetry_tetrahedron()
}
