% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{transform_molecule}
\alias{transform_molecule}
\title{Apply arbitratry 3D transformations to molecule3D objects}
\usage{
transform_molecule(x, transformation, ...)
}
\arguments{
\item{x}{a Molecule3D object}

\item{transformation}{A function that takes any length 3 numeric vector / named list of length 3
(\code{x}, \code{y}, \code{z}) and returns a structure (e.g., named numeric or list)
with elements \code{x}, \code{y}, and \code{z} in that order.}

\item{...}{additional arguments to transformation function}
}
\value{
A data frame with the same columns as \code{table}, but with the
\code{x}, \code{y}, and \code{z} coordinates replaced by the transformed values.
}
\description{
Applies a user-supplied transformation function to atoms in a Molecule3D object.
This lets you apply any function from the move package to a Molecule3D object.
}
\examples{
# Example: rotate points around Z-axis

path <- system.file(package="structures", "benzene.mol2")
molecule <- read_mol2(path)

rotate_z <- function(p) {
  angle <- pi / 4
  c(x = p["x"] * cos(angle) - p["y"] * sin(angle),
    y = p["x"] * sin(angle) + p["y"] * cos(angle),
    z = p["z"])
}

transform_molecule(molecule, rotate_z)

}
