% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{translate_molecule_by_vector}
\alias{translate_molecule_by_vector}
\title{Translate the molecule by a vector}
\usage{
translate_molecule_by_vector(x, vector)
}
\arguments{
\item{x}{A \code{Molecule3D} object.}

\item{vector}{Numeric length-3 vector \code{c(dx, dy, dz)}.}
}
\value{
A \code{Molecule3D} object translated by \code{vector}.
}
\description{
Adds a translation vector to all atom coordinates using
\code{\link{transform_molecule}}. The anchor is translated by the same vector,
keeping it consistent with the new coordinates.
}
\examples{

molecule
m <- translate_molecule_by_vector(m, c(1, 0, 0))

}
\seealso{
\code{\link{translate_molecule_to_origin}},
\code{\link{translate_molecule_to_position}},
\code{\link{set_anchor_by_position}}, \code{\link{set_anchor_by_atom}}
}
