% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{translate_molecule_in_direction}
\alias{translate_molecule_in_direction}
\title{Translate molecule in a given direction}
\usage{
translate_molecule_in_direction(x, direction, magnitude)
}
\arguments{
\item{x}{A \code{\link[=Molecule3D]{Molecule3D()}} object}

\item{direction}{Numeric vector giving the direction of translation (need not be unit length).}

\item{magnitude}{Numeric scalar giving the distance to move along \code{direction}.}
}
\value{
A \code{\link[=Molecule3D]{Molecule3D()}} object translated along \code{direction} by \code{magnitude} units
}
\description{
Translates all atom coordinates in a \code{\link{Molecule3D}} object
along a specified direction by a given distance.
}
\examples{
# m <- translate_molecule_in_direction(m, direction = c(1, 0, 0), magnitude = 2)
}
