% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{translate_molecule_to_origin}
\alias{translate_molecule_to_origin}
\title{Translate the molecule so its anchor is at the origin}
\usage{
translate_molecule_to_origin(x)
}
\arguments{
\item{x}{A \code{Molecule3D} object.}
}
\value{
A \code{Molecule3D} object translated so \code{x@anchor == c(0,0,0)}.
}
\description{
Moves all atom coordinates by a translation that places \code{x@anchor} at
\code{c(0, 0, 0)}. The anchor is updated accordingly via
\code{\link{transform_molecule}}.
}
\examples{
# m <- translate_molecule_to_origin(m)
}
\seealso{
\code{\link{translate_molecule_to_position}},
\code{\link{translate_molecule_by_vector}},
\code{\link{set_anchor_by_position}}, \code{\link{set_anchor_by_atom}}
}
