% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{valid_atom_types}
\alias{valid_atom_types}
\title{SYBYL bond types}
\usage{
valid_atom_types()
}
\value{
a character vector with all valid SYBYL atom types
}
\description{
SYBYL bond types
}
\examples{
valid_atom_types()
}
