% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{valid_bond_types}
\alias{valid_bond_types}
\title{SYBYL bond types}
\usage{
valid_bond_types()
}
\value{
a character vector with all valid SYBYL bond types
}
\description{
SYBYL bond types
}
\examples{
valid_bond_types()
}
