% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-SymmetryElements.R
\name{valid_symmetry_element_types}
\alias{valid_symmetry_element_types}
\title{Valid symmetry element types}
\usage{
valid_symmetry_element_types()
}
\value{
A character vector of valid symmetry element types.
}
\description{
Returns the set of recognised symmetry element type strings used by
\code{\link{SymmetryElement}} and its subclasses.
}
\examples{
valid_symmetry_element_types()

}
