% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parse.R
\name{write_mol2}
\alias{write_mol2}
\title{Write Mol2 file}
\usage{
write_mol2(molecule, path)
}
\arguments{
\item{molecule}{A \code{\link[=Molecule3D]{Molecule3D()}} object.}

\item{path}{path to mol2 file.}
}
\value{
invisibly returns NULL. This function is run for its side effects. .
}
\description{
Writes a \code{\link{Molecule3D}} object to disk in mol2 format.
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}

# Read mol2 file
mol2_path <- system.file("benzene.mol2", package = "structures")
mol <- read_mol2(mol2_path)

# Write minimal mol2 file
write_mol2(mol, "benzene.written.mol2")

\dontshow{
setwd(.old_wd)
}
}
