% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{compute_distance_between_atoms}
\alias{compute_distance_between_atoms}
\title{Compute distance between two points}
\usage{
compute_distance_between_atoms(x, eleno1, eleno2)
}
\arguments{
\item{x}{a Molecule3D object.}

\item{eleno1, eleno2}{atom IDs (eleno's) of the two atoms to compute distance between}
}
\value{
numeric euclidean distance between two points
}
\description{
Compute distance between two points
}
\examples{
path <- system.file(package="structures", "benzene.mol2")
molecule <- read_mol2(path)
compute_distance_between_atoms(molecule, 1, 2)
}
