% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Molecule3D.R
\name{locate_molecule_center}
\alias{locate_molecule_center}
\title{Locate geometrical center of molecule}
\usage{
locate_molecule_center(x)
}
\arguments{
\item{x}{a Molecule3D object}
}
\value{
a Molecule3D object
}
\description{
Locate geometrical center of all atoms in Molecule3D object. Can also just use center property.
Included as separate function to allow for tidy pipelines.
}
\examples{
path <- system.file(package="structures", "benzene.mol2")
molecule <- read_mol2(path)
locate_molecule_center(molecule)
}
