% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-SymmetryElements.R
\name{SymmetryElementCollection}
\alias{SymmetryElementCollection}
\title{SymmetryElementCollection}
\usage{
SymmetryElementCollection(
  symmetry_elements = list(),
  ids = integer(0),
  counts = (function (.data = list(), row.names = NULL) 
 {
     if (is.null(row.names))
    {
         list2DF(.data)
     }
     else {
         out <- list2DF(.data,
    length(row.names))
         attr(out, "row.names") <- row.names
         out
     }

    })()
)
}
\arguments{
\item{label}{Character scalar. Optional user-defined name/description.}

\item{type}{Character scalar. Read only. A string describing the element type. Must be one of  "Mirror Plane", "Centre of Inversion", "Improper Rotation Axis", "Proper Rotation Axis"}
}
\description{
A list of symmetry elements. Important for managing IDs of each element (ensuring they're unique)
Forces all symmetry elements to have label and type properties (see Arguments section).
}
